+++ src/cli/util.ts
@@ -38,6 +38,26 @@ export function passwordOption (description?: string) {
   )
 }
 
+export function currentAccountAddressOption () {
+  const storedConfig = config.getSync()
+  return new Option(
+    '-a, --account <accountAddress>',
+    'Sets the current account for the session or defaults to the account stored in the config'
+  )
+    .env('ACCOUNT_ADDRESS')
+    .argParser(async (address) => {
+      if (address === storedConfig.selectedAccount) return address
+      // Updated selected account in config with new address from this option
+      const newConfigUpdates = { selectedAccount: address }
+      await config.set({ ...storedConfig, ...newConfigUpdates })
+
+      return address
+    })
+    .hideHelp()
+    .default(storedConfig.selectedAccount)
+}
+
+
 export function aliasOrAddressOption () {
   return new Option(
     '-a, --address <aliasOrAddress>',