+++ src/common/base-command.ts
@@ -1,0 +1,12 @@
+import Entropy from "@entropyxyz/sdk";
+import { EntropyLogger } from "./logger";
+
+export abstract class BaseCommand {
+  protected logger: EntropyLogger
+  protected entropy: Entropy
+
+  constructor (entropy: Entropy, endpoint: string, flowContext: string) {
+    this.logger = new EntropyLogger(flowContext, endpoint)
+    this.entropy = entropy
+  }
+}