+++ src/transfer/types.ts
@@ -1,5 +1,6 @@
 // @ts-ignore
 import { Pair } from '@entropyxyz/sdk/keys'
+
 export interface TransferOptions { 
   from: Pair
   to: string