+++ src/tui.ts
@@ -1,15 +1,19 @@
 import inquirer from 'inquirer'
+import Entropy from '@entropyxyz/sdk'
 import * as config from './config'
 import * as flows from './flows'
 import { EntropyTuiOptions } from './types'
 import { logo } from './common/ascii'
 import { print } from './common/utils'
 import { EntropyLogger } from './common/logger'
+import { BalanceCommand } from './balance/command'
+import { TransferCommand } from './transfer/command'
+import { loadEntropy } from './cli'
 
 let shouldInit = true
 
 // tui = text user interface
-export default function tui (options: EntropyTuiOptions) {
+export default function tui (entropy: Entropy, options: EntropyTuiOptions) {
   const logger = new EntropyLogger('TUI', options.endpoint)
   console.clear()
   console.log(logo) // the Entropy logo
@@ -17,10 +21,11 @@ export default function tui (options: EntropyTuiOptions) {
 
   const choices = {
     'Manage Accounts': flows.manageAccounts,
-    'Balance': flows.checkBalance,
+    // leaving as a noop function until all flows are restructured
+    'Balance': () => {},
     'Register': flows.entropyRegister,
     'Sign': flows.sign,
-    'Transfer': flows.entropyTransfer,
+    'Transfer': () => {},
     // TODO: design programs in TUI (merge deploy+user programs)
     'Deploy Program': flows.devPrograms,
     'User Programs': flows.userPrograms,
@@ -36,10 +41,10 @@ export default function tui (options: EntropyTuiOptions) {
   // assign exit so its last
   Object.assign(choices, { 'Exit': async () => {} })
 
-  main(choices, options, logger)
+  main(entropy, choices, options, logger)
 }
 
-async function main (choices, options, logger: EntropyLogger) {
+async function main (entropy: Entropy, choices, options, logger: EntropyLogger) {
   if (shouldInit) {
     await config.init()
     shouldInit = false
@@ -52,6 +57,11 @@ async function main (choices, options, logger: EntropyLogger) {
   if (!storedConfig.selectedAccount && storedConfig.accounts.length) {
     await config.set({ ...storedConfig, ...{ selectedAccount: storedConfig.accounts[0].address } })
     storedConfig = await config.get()
+  }
+
+  // If the selected account changes within the TUI we need to reset the entropy instance being used
+  if (storedConfig.selectedAccount !== entropy.keyring.accounts.registration.address) {
+    entropy = await loadEntropy(storedConfig.selectedAccount, options.endpoint)
   }
 
   const answers = await inquirer.prompt([{
@@ -73,13 +83,42 @@ async function main (choices, options, logger: EntropyLogger) {
     console.error('There are currently no accounts available, please create or import your new account using the Manage Accounts feature')
   } else {
     logger.debug(answers)
-    const newConfigUpdates = await choices[answers.choice](storedConfig, options, logger)
-    if (typeof newConfigUpdates === 'string' && newConfigUpdates === 'exit') {
-      returnToMain = true
-    } else {
-      await config.set({ ...storedConfig, ...newConfigUpdates })
+    switch (answers.choice) {
+    case "Balance": {
+      try {
+        const balanceCommand = new BalanceCommand(entropy, options.endpoint)
+        const balanceString = await balanceCommand.getBalance(storedConfig.selectedAccount)
+        print(`Address ${storedConfig.selectedAccount} has a balance of: ${balanceString}`)
+      } catch (error) {
+        console.error('There was an error retrieving balance', error)
+      }
+      break;
     }
-    storedConfig = await config.get()
+    case "Transfer": {
+      try {
+        const transferCommand = new TransferCommand(entropy, options.endpoint)
+        const { amount, recipientAddress } = await transferCommand.askQuestions()
+        await transferCommand.sendTransfer(recipientAddress, amount)
+        print('')
+        print(`Transaction successful: Sent ${amount} to ${recipientAddress}`)
+        print('')
+        print('Press enter to return to main menu')
+      } catch (error) {
+        console.error('There was an error sending the transfer', error)
+      }
+      break
+    }
+    default: {
+      const newConfigUpdates = await choices[answers.choice](storedConfig, options, logger)
+      if (typeof newConfigUpdates === 'string' && newConfigUpdates === 'exit') {
+        returnToMain = true
+      } else {
+        await config.set({ ...storedConfig, ...newConfigUpdates })
+      }
+      storedConfig = await config.get()
+      break;
+    }
+    }
   }
 
   if (!returnToMain) {
@@ -90,7 +129,7 @@ async function main (choices, options, logger: EntropyLogger) {
     }]))
   }
 
-  if (returnToMain) main(choices, options, logger)
+  if (returnToMain) main(entropy, choices, options, logger)
   else {
     print('Have a nice day')
     process.exit()